requirements.txt
asgiref==3.5.2
Authlib==1.2.0
beautifulsoup4==4.11.2
blis==0.7.9
boto3==1.24.78
botocore==1.27.78
bs4==0.0.1
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
conllu==4.5.2
cryptography==39.0.1
cymem==2.0.7
Django==4.1.1
django-extensions==3.2.1
djangorestframework==3.14.0
en-ner-bionlp13cg-md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_ner_bionlp13cg_md-0.5.1.tar.gz
et-xmlfile==1.1.0
gensim==4.3.1
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-vision==3.4.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
Pillow==9.2.0
preshed==3.0.8
proto-plus==1.22.2
protobuf==4.22.1
psutil==5.9.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycparser==2.21
pydantic==1.10.6
pysbd==0.3.4
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.2
srsly==2.4.6
tesseract==0.1.3
thinc==8.1.9
threadpoolctl==3.1.0
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.2
urllib3==1.26.12
validate-email==1.3
wasabi==0.10.1
